remove unnecessary int64

+++ util.go
@@ -12,6 +12,6 @@ func SnowflakeTimestamp(ID string) (t time.Time, err error) {
 		return
 	}
 	timestamp := (i >> 22) + 1420070400000
-	t = time.Unix(0, timestamp*int64(1000000))
+	t = time.Unix(0, timestamp*1000000)
 	return
 }
